# docs/catbytes_api_swagger.yaml
openapi: 3.0.3
info:
  title: CatBytes Web Platform
  description: |-
    This is CatBytes Web Platform API specification.

    !
    Please note, this is an initial draft of the API, all payloads and response bodies are very likely to change.

    Endpoints marked with ✅ emoji are already implemented and live and  their specifications are up-to-date.

    Some useful links:
    - [CatBytes backend repository](https://github.com/catbytes-community/webplatform-backend)
    - [CatBytes API server](https://devapi.catbytes.io/)
  version: 1.0.0
...
servers:
  - url: https://devapi.catbytes.io/
...
paths:
  /applications:
    post:
      tags:
        - applications
      summary: ✅ Create new user application
      description: Send new user application's data for review. Whether video_link or video_filename is required.
      operationId: createApplication
      requestBody:
        description: Data user inserted on application
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Jane Doe
                about:
                  type: string
                  example: I'm passionate about C++ programming
                email:
                  type: string
                  example: jane@example.com
                video_link:
                  type: string
                  example: http://example.com/video/1
                  nullable: true
                video_filename:
                  type: string
                  example: videofile.mov
                  nullable: true
                discord_nickname:
                  type: string
                  example: sample_username
              oneOf:
                - required: [video_link]
                - required: [video_filename]
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 42
                  name: Jane Doe
                  about: I'm passionate about C++ programming
                  email: jane@example.com
                  video_link: http://example.com/video/1
                  video_filename: ''
                  discord_nickname: sample_username
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: null value in column "discord_nickname" of relation "applications" violates not-null constraint
        '409':
          description: Application with this email already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Application with this email already exists
    get:
      tags:
        - applications
      summary: ✅ Get all (active?) applications
      description: ''
      operationId: getApplications
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  applications:
                    type: array
                    items:
                      type: object
                      example:
                        - id: 5
                          name: jane doe
                          about: im cool
                          video_link: ''
                          discord_username: sample_username
                          created_at: '2024-11-23T16:37:04.115Z'
                          email: jane@example.com
                          video_filename: 'video1.mp4'
                        - id: 6
                          name: nina
                          about: i like cats
                          video_link: http://example.com/video/2
                          discord_username: nina_disc
                          created_at: '2024-11-20T16:37:04.115Z'
                          email: nina@example.com
                          video_filename: ''
        '403':
          description: User is not allowed to view applications list
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: You're not allowed to access this resource
  /applications/{application_id}:
    get:
      tags:
        - applications
      summary: ✅  Get membership application by id
      description: Returns application with extended information. Right now extended information only includes video_file - a link to uploaded video.
      operationId: getApplicationById
      parameters:
        - name: application_id
          in: path
          description: 'The id of application that needs to be fetched. Use 42 for testing. '
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Application found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  about:
                    type: string
                  email:
                    type: string
                  discord_nickname:
                    type: string
                  video_link:
                    type: string
                    nullable: true
                  video_filename:
                    type: string
                    nullable: true
                  video_file:
                    type: string
                    format: uri
                    description: Download URL for uploaded video
        '404':
          description: Application not found
        '500':
          description: Internal server error
      
    put:
      tags:
        - applications
      summary: ✅  Update application for membership
      description: This endpoint is used to either approve or reject an application to join the community (membership application)
      operationId: updateApplication
      parameters:
        - name: application_id
          in: path
          description: is of application that needs to be updated
          required: true
          schema:
            type: integer
      requestBody:
        description: If application is rejected, comment will be required and stored in database along with applications with rejected status). Application statuses is enum with options approved, rejected, pending
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  example: rejected
                comment:
                  type: string
                  example: it's a man!
                user_id:
                  type: integer
                  example: 1
      responses:
        '200':
          description: Successful operation
        '400':
          description: Unknown action
        '403':
          description: User is not allowed to edit applications
        '404':
          description: Application not found
...
  /mentors:
    post:
      tags:
        - mentors
      summary: ✅ Create a new mentor application
      description: |
        Creates a new mentor entity in 'pending' status. Available to all logged-in users. If user already has a mentor profile, creation is not allowed.
      operationId: createMentor
      requestBody:
        required: true
        description: Data provided by the user during application
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentorCreateRequest'
      responses:
        '201':
          description: Mentor application created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 42
        '400':
          description: User already has a mentor profile
        '403':
          description: User is not authenticated
    get:
      tags:
        - mentors
      summary: ✅  Get mentor list
      description: |
        Returns a list of mentor applications. When filtering by 'pending' or 'rejected', or not filtering at all, user must have elevated (admin) role. Returns user_id and some aggregated data from user profile
      operationId: getMentors
      parameters:
        - name: status
          in: query
          description: Status of mentors to retrieve (e.g. active, pending, rejected)
          required: false
          schema:
            type: string
            enum:
              - active
              - pending
              - rejected
              - inactive
      responses:
        '200':
          description: List of mentors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorListResponse'
        '403':
          description: User is not allowed to view this list
  /mentors/{mentor_id}:
    get:
      tags:
        - mentors
      summary: ✅ Get mentor by ID
      description: |
        Returns full mentor data if user is logged in. Pending/rejected mentors require admin role.
      operationId: getMentorById
      parameters:
        - name: mentor_id
          in: path
          required: true
          description: ID of the mentor to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: Mentor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorDetails'
        '403':
          description: Access denied
        '404':
          description: Mentor not found
    put:
      tags:
        - mentors
      summary: ✅ Update mentor profile
      description: |
        Only the user who owns this mentor card can update it. Only selected fields can be changed.
      operationId: updateMentor
      parameters:
        - name: mentor_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentorUpdateRequest'
      responses:
        '200':
          description: Mentor profile updated
        '403':
          description: Not authorized
        '404':
          description: Mentor not found
    patch:
      tags:
        - mentors
      summary: ✅ Update mentor status
      description: |
        Available to users with elevated role (admin) only or to user owning the mentor card.
      operationId: patchMentorStatus
      parameters:
        - name: mentor_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - active
                    - rejected
                    - inactive
      responses:
        '200':
          description: Status updated
        '403':
          description: Not authorized
        '404':
          description: Mentor not found
    delete:
      tags:
        - mentors
      summary: ✅ Delete mentor profile
      description: |
        Only the owner can delete their mentor card. Deletes the entity and removes mentor role from user.
      operationId: deleteMentor
      parameters:
        - name: mentor_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Mentor deleted
        '403':
          description: Not authorized
        '404':
          description: Mentor not found
...
components:
  schemas:
    MentorCreateRequest:
      type: object
      properties:
        contact:
          type: string
        about:
          type: string
        tags:
          type: array
          items:
            type: string
    MentorUpdateRequest:
      type: object
      properties:
        contact:
          type: string
        about:
          type: string
        tags:
          type: array
          items:
            type: string
    MentorListResponse:
      type: object
      properties:
        mentors:
          type: array
          items:
            $ref: '#/components/schemas/MentorSummary'
    MentorSummary:
      type: object
      properties:
        mentor_id:
          type: integer
        user_id:
          type: integer
        about:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        name:
          type: string
        languages:
          type: array
          items:
            type: string
        img_link:
          type: string
    MentorDetails:
      allOf:
        - $ref: '#/components/schemas/MentorSummary'
        - type: object
          properties:
            contact:
              type: string
            discord_nickname:
              type: string